requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
blinker==1.8.2
browser-cookie3==0.19.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==5.5
deprecation==2.1.0
distlib==0.3.8
et-xmlfile==1.1.0
exceptiongroup==1.2.1
eyed3==0.9.7
filelock==3.12.4
filetype==1.2.0
Flask==3.0.3
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.9.2
gino==1.0.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
huggingface-hub==0.22.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
liboml==0.0.2
lightgbm==4.1.0
lz4==4.3.3
macholib==1.16.3
magic-filter==1.0.12
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.23.6
openpyxl==3.1.2
packaging==23.2
pango==0.0.1
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2==2.9.9
pyaes==1.6.1
pyasn1==0.5.1
pycryptodomex==3.21.0
pydantic==2.7.1
pydantic_core==2.18.2
pydotenv==0.0.7
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.1.1
pyproject-toml==0.0.10
pysam==0.22.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.11.3
six==1.16.0
sklearn==0.0.post9
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.3.24
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.1.0
tqdm==4.66.2
typing_extensions==4.8.0
urllib3==2.2.0
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.25.0
watchfiles==0.24.0
websockets==14.1
Werkzeug==3.0.3
yandex-music-downloader @ git+https://github.com/llistochek/yandex-music-downloader@4de447950b3683b111e24aa63beed68b021b9024
yarl==1.9.4
zipp==3.17.0
